\name{ryThesis-package}
\alias{ryThesis-package}
\alias{ryThesis}
\docType{package}
\title{
The functions from my Honors Thesis
}
\description{
This package contains all the R functions I wrote for my honors thesis project.
}
\details{
\tabular{ll}{
Package: \tab ryThesis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-08\cr
License: \tab GPL 2\cr
}
}
\author{
Ryan Himmelwright

Maintainer: Ryan Himmelwright to <ryan.himmelwright@gmail.com>
}
\references{
To add.
}
\examples{
print("hi")
}
